wip: combine loupedeck implementations into one
